Moved SessionLog tests to dedicated class
